refactor: Rename djb2Hash file for consistency
Renamed djb2Hash.ts to djb2-hash.ts to maintain a consistent naming convention throughout the codebase. This change helps to improve code readability and organization.
